+++ HexInvestigator/Utils.cpp
@@ -1,6 +1,6 @@
 #include "Utils.h"
 
-void Utils::WriteValue(HANDLE procHandle, unsigned long long* address, void* newValue, int size)
+void Utils::WriteValue(HANDLE procHandle, uintptr_t* address, void* newValue, int size)
 {
 	DWORD oldProtect;
 	VirtualProtectEx(procHandle, address, size, PAGE_EXECUTE_READWRITE, &oldProtect);
